fix android build error
At present, I have not added DataAnalyst to Android, so there is no entry point for MemorySniffer-related functions, and it cannot be used on Android.

+++ src/core/memory/brace_script/brace_object.inl
@@ -4915,7 +4915,7 @@ namespace BraceScriptInterpreter
                 int stat = m_SelectStats[vix];
                 auto& itInfo = m_Iterators[vix++];
                 if (stat == STAT_METHOD_AVG) {
-                    cv.NumVal /= group.size();
+                    cv.NumVal /= static_cast<double>(group.size());
                 }
                 if (Brace::IsStringType(itInfo.Type))
                     Brace::VarSetString(lvars, itInfo.VarIndex, cv.StrVal);
